fixed ind

+++ libc/inc/driver.h
@@ -8,7 +8,7 @@
 /***** PORT ACCESS *****/
 uint8_t inb(uint16_t port);
 uint16_t inw(uint16_t port);
-uint16_t ind(uint16_t port);
+uint32_t ind(uint16_t port);
 
 void outb(uint16_t port, uint8_t value);
 void outw(uint16_t port, uint16_t value);
